requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.6,<4.7
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<1.1
google-genai>=1.0,<2.0
